pyproject.toml
# pyproject.toml
[project]
name = "byu-flagellar-detection"
version = "1.0.0"
dependencies = [
    "plotly>=5.18.0",
    "pillow>=10.0.0",       # PIL
    "seaborn>=0.13.0",
    "matplotlib>=3.7.0",
    "ultralytics>=8.1.0",   # YOLOv8
    "pyyaml>=6.0.0",        # yaml
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",  # sklearn
    "torch>=2.0.1",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0", # cv2
    "tqdm>=4.66.0",
    "scipy>=1.11.0"
]

[project.optional-dependencies]
dev = [
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.0.0",
    "pytest>=7.4.0"
]